refactor test_circuit into submodules
