+++ tests/test_tensor/test_circuit/_helpers.py
@@ -1,0 +1,20 @@
+def rand_reg_graph(reg, n, seed=None):
+    import networkx as nx
+
+    G = nx.random_regular_graph(reg, n, seed=seed)
+    return G
+
+
+def graph_to_qsim(G, gamma0=-0.743043, beta0=0.754082):
+    n = G.number_of_nodes()
+
+    # add all the gates
+    circ = f"{n}\n"
+    for i in range(n):
+        circ += f"H {i}\n"
+    for i, j in G.edges:
+        circ += f"Rzz {gamma0} {i} {j}\n"
+    for i in range(n):
+        circ += f"Rx {beta0} {i}\n"
+
+    return circ
